Starts introduction areas. Reformmated starting pos from design to start lower. See description.
My original design started the player off downhill and moving uphill toward the outer castle walls. I wanted the walls to come into view and feel massive by purposelly making the player come up to it from a down angle. But in practice I found that in order to get that effect the hill had to be very steep, or very far away from the player. I didn't like this so I instead started the player off at the bottom of a small bunker and moving up the staircase. I'll append the storyline of him rising from hell, or something to that effect. The staircase is tiny compared to the opening grassy area, so it adds to the contrast of a small then large area. The staircase with a small open door also gives me the effect I wanted of the castle walls coming into view, and the steep downward angle adds to this effect. I think this will be interesting from the player's perspective. I plan to make the starting area dark and the outer area light as well to add to the contrast.