Adds initial teleporter puzzle. Still needs end.
As stated in the top-down, this puzzle works  by subetling changing the environment when the player progresses. This will be dictated by differing lights and a more macabre atmpsohere. Unforunttalye adding one light source means I gotta add lights for the whole map which I don't want to focus on right now. And I don't want to focus on textures just yet either, but I needed to make some kind of identifier of the diffrent rooms using the prototype textures, so I used darker shades of the prototype texture.